refactoring

+++ src/gprogression.rs
@@ -1,13 +1,13 @@
 use std::error::Error;
 
-pub fn sum(b1: f64, q: f64, n: usize) -> Result<f64, Box<dyn Error>> {
+pub fn sum(b1: f32, q: f32, n: usize) -> Result<f32, Box<dyn Error>> {
     //! sum = b1 * (q^n - 1) / (q - 1)
 
     if q == 1.0 { return Err("Argument 'q' must not be 1.0".into()) }
-    Ok(b1 * (q.powf(n as f64) - 1.0) / (q - 1.0))
+    Ok(b1 * (q.powf(n as f32) - 1.0) / (q - 1.0))
 }
 
-pub fn get_n(b1: f64, q: f64, sum: f64) -> usize {
+pub fn get_n(b1: f32, q: f32, sum: f32) -> usize {
     //! 'n' - number of iterations will be spent before 'sum' is achieved
 
     let mut n = Box::<usize>::new(1);
@@ -17,7 +17,7 @@ pub fn get_n(b1: f64, q: f64, sum: f64) -> usize {
     *n
 }
 
-pub fn get_b_n(b_k: f64, q: f64, k: usize, n: usize) -> Result<f64, Box<dyn Error>> {
+pub fn get_b_n(b_k: f32, q: f32, k: usize, n: usize) -> Result<f32, Box<dyn Error>> {
     //! b_n = b_k * q^(n - k) , n > k
     //!
     //! b_n = b_k / q^(k - n) , n < k
@@ -32,12 +32,12 @@ pub fn get_b_n(b_k: f64, q: f64, k: usize, n: usize) -> Result<f64, Box<dyn Erro
 
     if q == 0.0 { return Err("Argument 'q' must not be 0".into()) }
 
-    if n > k { return Ok(b_k * q.powf((n - k) as f64)) }
-    Ok(b_k / q.powf((k - n) as f64))
+    if n > k { return Ok(b_k * q.powf((n - k) as f32)) }
+    Ok(b_k / q.powf((k - n) as f32))
 }
 
 #[allow(dead_code)]
-pub fn get_q(b_k: f64, b_n: f64, k: usize, n: usize) -> Result<f64, Box<dyn Error>> {
+pub fn get_q(b_k: f32, b_n: f32, k: usize, n: usize) -> Result<f32, Box<dyn Error>> {
     //! q = (b_n / b_k)^(1 / (n - k)) , n > k
     //!
     //! q = (b_k / b_n)^(1 / (k - n)) , n < k
@@ -54,8 +54,8 @@ pub fn get_q(b_k: f64, b_n: f64, k: usize, n: usize) -> Result<f64, Box<dyn Erro
 
     if n > k {
         if b_k == 0.0 { return Err("Argument 'b_k' must not be 0".into()) }
-        return Ok((b_n / b_k).powf(1.0 / (n - k) as f64))
+        return Ok((b_n / b_k).powf(1.0 / (n - k) as f32))
     }
     if b_n == 0.0 { return Err("Argument 'b_n' must not be 0".into()) }
-    Ok((b_k / b_n).powf(1.0 / (k - n) as f64))
+    Ok((b_k / b_n).powf(1.0 / (k - n) as f32))
 }
