+++ src/top.rs
@@ -1,7 +1,7 @@
 use std::error::Error;
 use crate::common::factorial;
 
-pub fn binominal_coefficient(m: usize, n: usize) -> Result<f64, Box<dyn Error>> {
+pub fn binominal_coefficient(m: usize, n: usize) -> Result<f32, Box<dyn Error>> {
     //! C_n^m = n! / m! * (n - m)! , m < n
 
     if m > n { return Err("Argument 'n' must be greater then 'm'.".into()); }
@@ -12,11 +12,11 @@ pub fn binominal_coefficient(m: usize, n: usize) -> Result<f64, Box<dyn Error>> 
     for i in m+1..=n {
         *opt_n *= i as u128;
     }
-    Ok(*opt_n as f64 / factorial(n - m) as f64 )
+    Ok(*opt_n as f32 / factorial(n - m) as f32 )
 }
 
 
-pub fn bernoulli(m: usize, n: usize, p: f64, q: f64) -> Result<f64, Box<dyn Error>> {
+pub fn bernoulli(m: usize, n: usize, p: f32, q: f32) -> Result<f32, Box<dyn Error>> {
     //! P_n^m = C_n^m * p^m * q^(n-m) , m < n
     //!
     //! m - required number of tests
@@ -28,7 +28,7 @@ pub fn bernoulli(m: usize, n: usize, p: f64, q: f64) -> Result<f64, Box<dyn Erro
     //! q - probability that the event will not occur in each test
 
     match crate::top::binominal_coefficient(m, n) {
-        Ok(c) => Ok(c * p.powf(m as f64) * q.powf((n - m) as f64)),
+        Ok(c) => Ok(c * p.powf(m as f32) * q.powf((n - m) as f32)),
         Err(e) => Err(e)
     }
 }
