+++ src/utils.rs
@@ -1,4 +1,4 @@
-pub fn is_equal(a: &f64, b: &f64, precision: f64) -> bool {
+pub fn is_equal(a: &f32, b: &f32, precision: f32) -> bool {
     if (a - b).abs() <= precision { return true }
     false
 }
