+++ tests/test_top.rs
@@ -5,7 +5,7 @@ mod test_theory_of_probability {
     fn test_binominal_coefficient1_pos() {
         // C_n^m = n! / m! * (n - m)!
         let res = binominal_coefficient(4, 5).unwrap();
-        let res_test: f64 = (2.0 * 3.0 * 4.0 * 5.0) / (2.0 * 3.0 * 4.0) * (5.0 - 4.0);
+        let res_test: f32 = (2.0 * 3.0 * 4.0 * 5.0) / (2.0 * 3.0 * 4.0) * (5.0 - 4.0);
         assert_eq!(res, res_test);
     }
 
@@ -30,10 +30,10 @@ mod test_theory_of_probability {
     #[test]
     fn test_bernoulli1_pos() {
         // P_n^m = C_n^m * p^m * q^(n-m)
-        let q: f64 = 0.5;
-        let p: f64 = 0.5;
+        let q: f32 = 0.5;
+        let p: f32 = 0.5;
         let res = bernoulli(7, 15, p, q).unwrap();
-        let res_test: f64 = binominal_coefficient(7, 15).unwrap() * p.powf(7.0) * q.powf(15.0-7.0);
+        let res_test: f32 = binominal_coefficient(7, 15).unwrap() * p.powf(7.0) * q.powf(15.0-7.0);
         assert_eq!(res, res_test);
     }
